chore(alerts): find and remove orphaned files (#13498)
* chore: clean up unused files (per Knip)

* chore: rename NewSavedSearchAlert* components

* chore: rename NewSavedSearchAlert* file

* chore: rename NewSavedSearchAlert* fragments
